Refine ThunderAgent Harbor 32B recipe
