+++ examples/train/thunder_agent/scripts/r2egym_32b/cleanup_docker.sh
