+++ examples/train/thunder_agent/scripts/r2egym_32b/prepull_images.py
