+++ examples/train/thunder_agent/skyrl_integration/__init__.py
@@ -1,11 +1,11 @@
 """SkyRL-specific ThunderAgent integration helpers."""
 
-from .generator import ThunderAgentSkyRLGymGenerator
+from .generator import ThunderAgentHarborGenerator
 from .remote_inference_client import ThunderAgentRemoteInferenceClient
 from .router import ThunderAgentRouter
 
 __all__ = [
-    "ThunderAgentSkyRLGymGenerator",
+    "ThunderAgentHarborGenerator",
     "ThunderAgentRemoteInferenceClient",
     "ThunderAgentRouter",
 ]
