+++ examples/train/thunder_agent/skyrl_integration/runtime_setup.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import os
 
-
 _PATCHED = False
 
 
@@ -21,8 +20,12 @@ def patch_mini_swe_agent_environment() -> None:
 
     def create_run_agent_commands(self, instruction):
         commands = original(self, instruction)
-        shared_home = os.environ.get("HARBOR_SHARED_MINI_SWE_TOOL_ENV_HOME", "/tmp/harbor-mini-swe-home")
-        uv_cache_dir = os.environ.get("HARBOR_SHARED_UV_CACHE_ENV_DIR", "/harbor-shared/uv-cache")
+        shared_home = os.environ.get(
+            "HARBOR_SHARED_MINI_SWE_TOOL_ENV_HOME", "/tmp/harbor-mini-swe-home"
+        )
+        uv_cache_dir = os.environ.get(
+            "HARBOR_SHARED_UV_CACHE_ENV_DIR", "/harbor-shared/uv-cache"
+        )
         shared_bin = f"{shared_home}/.local/bin"
         package_ref = os.environ.get(
             "HARBOR_MINI_SWE_AGENT_PACKAGE",
@@ -31,7 +34,9 @@ def patch_mini_swe_agent_environment() -> None:
 
         for command in commands:
             env = dict(command.env or {})
-            env["PATH"] = f"{shared_bin}:{env.get('PATH') or os.environ.get('PATH', '')}"
+            env["PATH"] = (
+                f"{shared_bin}:{env.get('PATH') or os.environ.get('PATH', '')}"
+            )
             env.setdefault("UV_LINK_MODE", "copy")
             command.env = env
             command.command = (
@@ -39,7 +44,7 @@ def patch_mini_swe_agent_environment() -> None:
                 f"export UV_CACHE_DIR={uv_cache_dir}; "
                 f"export PATH={shared_bin}:$PATH; "
                 "if ! mini --help >/dev/null 2>&1; then "
-                "mkdir -p \"$HOME\" \"$UV_CACHE_DIR\"; "
+                'mkdir -p "$HOME" "$UV_CACHE_DIR"; '
                 f"uv tool install --force --python /usr/bin/python3.10 --cache-dir {uv_cache_dir} {package_ref}; "
                 "hash -r; "
                 "fi; "
